refactor: Update register field types and enhance protocol handling
- Replaced RegField08 with RegFieldU08 across multiple driver files for consistency in unsigned types.
- Updated platformio.ini to enable native environment configuration.
- Modified PortHandler class to change response delay from uint16_t to float for improved precision.
- Enhanced RegMap class to utilize unsigned types in register field operations, ensuring better type safety.
- Introduced TimeoutLimiter class for managing timeout conditions in servo operations, improving error handling and response management.
- Improved documentation to reflect changes in data types and new functionalities.

+++ lib/hortor/src/protocol/port_handler.h
@@ -62,15 +62,15 @@ class PortHandler {
 
   /**
    * @brief 设置响应延迟
-   * @param response_delay 响应延迟时间（毫秒）
+   * @param response_delay 响应延迟时间（秒）
    */
-  void SetResponseDelay(const uint16_t response_delay) {
+  void SetResponseDelay(const float response_delay) {
     response_delay_ = response_delay;
   }
 
  protected:
   StatusPacket status_packet_{};  // 状态包缓冲区
-  uint16_t response_delay_ = 0;   // 响应延迟（毫秒）
+  float response_delay_ = 0.0f;   // 响应延迟（秒）
 };
 
 }  // namespace hortor::protocol
